Start splitting tests by functionality.
